+++ lib/features/dataset/text_field.dart
@@ -93,7 +93,9 @@ class DatasetTextFieldState extends ConsumerState<DatasetTextField> {
       child: MarkdownTooltip(
         message: '''
 
-        **Path:** You can paste or type the path to a file containing your
+        **Path**
+
+        You can paste or type the path to a file containing your
         dataset. It is expected to be a **csv** or **txt** file, or the name of
         a package dataset, like rattle::wattle.