+++ lib/features/wordcloud/tooltips.dart
@@ -29,13 +29,17 @@ library;
 
 const textMineToolTip = '''
 
-**Text Mine:** Tap here to build the analysis of the text document(s).
+**Text Mine**
+
+Tap here to build the analysis of the text document(s).
 
 ''';
 
 const randomOrderToolTip = '''
 
-**Random Order:** Tick the checkbox to have the word cloud generated with a
+**Random Order**
+
+Tick the checkbox to have the word cloud generated with a
 random ordering of the words.  Otherwise (the default) the most frequent words
 are centered and then other words are drawn in decreasing frequency as we
 progress to the edg of the picture.
@@ -44,14 +48,18 @@ progress to the edg of the picture.
 
 const saveCsvToolTip = '''
 
-**Save to CSV:** Tap here to save the generated document term matrix to a CSV
+**Save to CSV**
+
+Tap here to save the generated document term matrix to a CSV
 file, one row for each document and a column for each term.
 
 ''';
 
 const stemToolTip = '''
 
-**Stem:**. Enable this to reduces words to their base or root form.  Two
+**Stem**
+
+Enable this to reduces words to their base or root form.  Two
 different words, when stemmed, can become the same and so can reduce unnecessary
 clutter in the wordcloud.
 
@@ -59,20 +67,26 @@ clutter in the wordcloud.
 
 const punctuationToolTip = '''
 
-**Punctuation:** Extraneous data can be removed including various punctuation.
+**Punctuation**
+
+Extraneous data can be removed including various punctuation.
 
 ''';
 
 const stopwordsToolTip = '''
 
-**Stopwords:** Remove common language words. The words removed depend on the
+**Stopwords**
+
+Remove common language words. The words removed depend on the
 chosen language.
 
 ''';
 
 const languageToolTip = '''
 
-**Language:** The stopwords removed will depend on the language. Select the
+**Language**
+
+The stopwords removed will depend on the language. Select the
 language of choice here to filter out common stopwords.
 
 The stopwords come from